+++ client/src/main/java/Class139.java
@@ -50,7 +50,7 @@ public final class Class139 {
 
 	@OriginalMember(owner = "client!gba", name = "a", descriptor = "(IZ)Lclient!ih;")
 	public Class2_Sub2_Sub12 method2950(@OriginalArg(0) int arg0) {
-		@Pc(11) Class2_Sub2_Sub12 local11 = (Class2_Sub2_Sub12) this.aClass82_70.method2156((long) arg0);
+		@Pc(11) Class2_Sub2_Sub12 local11 = (Class2_Sub2_Sub12) this.aClass82_70.method2156(arg0);
 		if (local11 != null) {
 			return local11;
 		}
@@ -68,7 +68,7 @@ public final class Class139 {
 		if (arg0 >= 32768) {
 			local11.method3902();
 		}
-		this.aClass82_70.method2150(local11, (long) arg0);
+		this.aClass82_70.method2150(local11, arg0);
 		return local11;
 	}
 }
