+++ client/src/main/java/Class147.java
@@ -9,7 +9,7 @@ public final class Class147 implements Interface23 {
 	@OriginalMember(owner = "client!gl", name = "a", descriptor = "(II)I")
 	@Override
 	public int method6872(@OriginalArg(0) int arg0) {
-		@Pc(18) Class2_Sub38 local18 = (Class2_Sub38) Static25.aClass28_4.method738((long) arg0);
+		@Pc(18) Class2_Sub38 local18 = (Class2_Sub38) Static25.aClass28_4.method738(arg0);
 		return local18 == null ? Static34.aClass306_1.method6872(arg0) : local18.anInt6399;
 	}
